requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.5
pandas==2.3.2
pyyaml==6.0.2
redis==6.4.0
celery==5.5.3
httpx==0.28.1
geoalchemy2==0.17.1
shapely==2.0.7
pytest==8.4.1
pytest-asyncio==1.1.0
